refactor(cli): swap old api rest caching with undici caching
