+++ packages/@expo/cli/src/api/rest/cache/__tests__/wrapFetchWithCache.test.ts
@@ -15,10 +15,11 @@ beforeEach(() => vol.fromJSON({ '/test/.gitkeep': '' }));
 
 it('returns cached response for get request', async () => {
   const server = jest.fn(() => ({ get: true }));
-  const scope = nock('http://doesnot.exist').get('/test/get').reply(200, server);
+
+  nock('http://expo.test').get('/get').reply(200, server);
 
   function fetchAction() {
-    return fetchWithCache('http://doesnot.exist/test/get').then((res) => res.json());
+    return fetchWithCache('http://expo.test/get').then((res) => res.json());
   }
 
   const response = await fetchAction();
@@ -27,18 +28,17 @@ it('returns cached response for get request', async () => {
   expect(server).toHaveBeenCalledTimes(1);
   expect(response).toEqual(server());
   expect(response).toEqual(cachedResponse);
-
-  expect(scope.isDone()).toBe(true);
 });
 
 it('returns cached response for post request with json body', async () => {
   const server = jest.fn(() => ({ post: 'json-body' }));
-  const scope = nock('http://doesnot.exist')
-    .post('/test/post', { test: 'json-body' })
+
+  nock('http://expo.test')
+    .post('/post', (body) => body.test === 'json-body')
     .reply(201, server);
 
   function fetchAction() {
-    return fetchWithCache('http://doesnot.exist/test/post', {
+    return fetchWithCache('http://expo.test/post', {
       method: 'POST',
       body: JSON.stringify({ test: 'json-body' }),
     }).then((response) => response.json());
@@ -50,21 +50,20 @@ it('returns cached response for post request with json body', async () => {
   expect(server).toHaveBeenCalledTimes(1);
   expect(response).toEqual(server());
   expect(response).toEqual(cachedResponse);
-
-  expect(scope.isDone()).toBe(true);
 });
 
 it('returns cached response for post request formdata body', async () => {
   const server = jest.fn(() => ({ post: 'formdata-body' }));
-  const scope = nock('http://doesnot.exist')
-    .post('/test/post', (body) => body.includes('formdata-body'))
+
+  nock('http://expo.test')
+    .post('/post', (body) => body.includes('formdata-body'))
     .reply(201, server);
 
   function fetchAction() {
     const body = new FormData();
     body.append('test', 'formdata-body');
 
-    const request = fetchWithCache('http://doesnot.exist/test/post', { body, method: 'POST' });
+    const request = fetchWithCache('http://expo.test/post', { body, method: 'POST' });
     return request.then((response) => response.json());
   }
 
@@ -74,6 +73,22 @@ it('returns cached response for post request formdata body', async () => {
   expect(server).toHaveBeenCalledTimes(1);
   expect(response).toEqual(server());
   expect(response).toEqual(cachedResponse);
+});
 
-  expect(scope.isDone()).toBe(true);
+// NOTE(cedric): error-responses aren't properly handled in nock@14.0.0-beta.7, re-enable once fixed
+xit('does not cache failed respose for get request', async () => {
+  const server = jest.fn(() => ({ error: 'not found' }));
+
+  nock('http://expo.test').get('/error/get').reply(404, server);
+
+  function fetchAction() {
+    return fetchWithCache('http://expo.test/error/get').then((res) => res.json());
+  }
+
+  const response = await fetchAction();
+  const cachedResponse = await fetchAction();
+
+  expect(server).toHaveBeenCalledTimes(2);
+  expect(response).toEqual(server());
+  expect(response).toEqual(cachedResponse);
 });
